Move provider code into its own files
